+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 v0.13 includes:
 * Removed prettier render to make html code look prettier, as a move to migrate that process as an input option.
 * Added namespace template compilation.
+* Fixed error when putting path without an ending / or \ for output-path input
 
 v0.12 includes:
 * Updated QuickTypeInfo to reflect Sharp Checker.